assert/cmp: add Regexp()
Checking if a string matches a regexp is useful. Currently it
can be implemented with something like

	assert.Assert(t, regexp.MustCompile("^Got: [0-9]+$").MatchString(str))

which looks borderline ugly and does not provide the value of str
in case of mismatch.

With the new functionality, one can use

	assert.Assert(t, is.Regexp("^Got: [0-9]+$", str))

or, in case they'd prefer a precompiled regex:

	r, err := regexp.Compile("^Got: [0-9]+$")
	assert.NilError(err)
	...
	assert.Assert(t, is.Regexp(r, str))

Currently the first parameter can only be a string or a precompiled regexp,
and the second parameter can only be a string. I see no profit in adding
any other types, as any type that can be easily converted to string
can be converted by the caller, and any other type probably does
not make sense.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>