fix TestFromDirSymlink on Windows due to missing drive-letter
This test was failing, because it expected the drive-letter to always
be returned, which wasn't the case;

    === FAIL: fs TestFromDirSymlink (0.01s)
        ops_test.go:55: assertion failed: directory C:\Users\circleci\AppData\Local\Temp\test-from-dir-2260318643 does not match expected:
            \a\b\3
              target: expected C:\some\inexistent\link got \some\inexistent\link

This patch removes the special-case from d44953da5fd92b6e954684d9fbc989a5be762bec,
which unconditionally added the drive-letter. Instead, we only change
the check to use the expected directory-separator.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

test

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>